Add 3-set_bit.c

+++ 0x14-bit_manipulation/3-set_bit.c
@@ -11,7 +11,7 @@
 
 int set_bit(unsigned long int *n, unsigned int index)
 {
-	unsigned long int bin_num = (sizeof(*(n)) * 11);
+	size_t bin_num = (sizeof(*(n)) * 11);
 
 	if (index < bin_num)
 	{
